refactor: split trace command helpers
