+++ src/commands/trace/output.rs
@@ -1,0 +1,352 @@
+use std::collections::{BTreeMap, BTreeSet};
+use std::path::{Path, PathBuf};
+
+use serde::Deserialize;
+
+use homeboy::extension::trace as extension_trace;
+use homeboy::extension::trace::TraceCommandOutput;
+
+use super::TraceArgs;
+use crate::commands::CmdResult;
+
+#[derive(Deserialize)]
+pub(super) struct TraceAggregateInput {
+    pub(super) component: Option<String>,
+    pub(super) scenario_id: Option<String>,
+    pub(super) spans: Vec<TraceAggregateSpanInput>,
+}
+
+#[derive(Deserialize)]
+struct TraceAggregateEnvelopeInput {
+    data: TraceAggregateInput,
+}
+
+#[derive(Deserialize)]
+pub(super) struct TraceAggregateSpanInput {
+    pub(super) id: String,
+    pub(super) n: usize,
+    pub(super) median_ms: Option<u64>,
+    pub(super) avg_ms: Option<f64>,
+    pub(super) failures: usize,
+}
+
+pub(super) fn run_compare(args: TraceArgs) -> CmdResult<TraceCommandOutput> {
+    let before_path = PathBuf::from(&args.scenario);
+    let Some(after_path) = args.compare_after else {
+        return Err(homeboy::Error::validation_invalid_argument(
+            "AFTER_JSON",
+            "trace compare requires before and after aggregate JSON files",
+            None,
+            None,
+        ));
+    };
+
+    let before = read_trace_aggregate(&before_path)?;
+    let after = read_trace_aggregate(&after_path)?;
+    let output = compare_trace_aggregates(&before_path, before, &after_path, after);
+    Ok((TraceCommandOutput::Compare(output), 0))
+}
+
+fn read_trace_aggregate(path: &Path) -> homeboy::Result<TraceAggregateInput> {
+    let content = std::fs::read_to_string(path).map_err(|err| {
+        homeboy::Error::internal_io(
+            format!("Failed to read trace aggregate {}: {}", path.display(), err),
+            Some("trace.compare.read".to_string()),
+        )
+    })?;
+    parse_trace_aggregate_input(&content).map_err(|err| {
+        homeboy::Error::internal_json(
+            err.to_string(),
+            Some(format!("parse trace aggregate {}", path.display())),
+        )
+    })
+}
+
+pub(super) fn parse_trace_aggregate_input(
+    content: &str,
+) -> serde_json::Result<TraceAggregateInput> {
+    match serde_json::from_str::<TraceAggregateInput>(content) {
+        Ok(input) => Ok(input),
+        Err(direct_error) => serde_json::from_str::<TraceAggregateEnvelopeInput>(content)
+            .map(|envelope| envelope.data)
+            .map_err(|_| direct_error),
+    }
+}
+
+pub(super) fn compare_trace_aggregates(
+    before_path: &Path,
+    before: TraceAggregateInput,
+    after_path: &Path,
+    after: TraceAggregateInput,
+) -> extension_trace::TraceCompareOutput {
+    let before_spans = before
+        .spans
+        .into_iter()
+        .map(|span| (span.id.clone(), span))
+        .collect::<BTreeMap<_, _>>();
+    let after_spans = after
+        .spans
+        .into_iter()
+        .map(|span| (span.id.clone(), span))
+        .collect::<BTreeMap<_, _>>();
+    let span_ids = before_spans
+        .keys()
+        .chain(after_spans.keys())
+        .cloned()
+        .collect::<BTreeSet<_>>();
+
+    let spans = span_ids
+        .into_iter()
+        .map(|id| {
+            let before_span = before_spans.get(&id);
+            let after_span = after_spans.get(&id);
+            let before_median = before_span.and_then(|span| span.median_ms);
+            let after_median = after_span.and_then(|span| span.median_ms);
+            let before_avg = before_span.and_then(|span| span.avg_ms);
+            let after_avg = after_span.and_then(|span| span.avg_ms);
+
+            extension_trace::TraceCompareSpanOutput {
+                id,
+                before_n: before_span.map(|span| span.n),
+                after_n: after_span.map(|span| span.n),
+                before_median_ms: before_median,
+                after_median_ms: after_median,
+                median_delta_ms: option_delta_i64(before_median, after_median),
+                median_delta_percent: option_percent_delta(
+                    before_median.map(|value| value as f64),
+                    after_median.map(|value| value as f64),
+                ),
+                before_avg_ms: before_avg,
+                after_avg_ms: after_avg,
+                avg_delta_ms: option_delta_f64(before_avg, after_avg),
+                avg_delta_percent: option_percent_delta(before_avg, after_avg),
+                before_failures: before_span.map(|span| span.failures),
+                after_failures: after_span.map(|span| span.failures),
+            }
+        })
+        .collect::<Vec<_>>();
+
+    extension_trace::TraceCompareOutput {
+        command: "trace.compare.spans",
+        before_path: before_path.display().to_string(),
+        after_path: after_path.display().to_string(),
+        before_component: before.component,
+        after_component: after.component,
+        before_scenario_id: before.scenario_id,
+        after_scenario_id: after.scenario_id,
+        span_count: spans.len(),
+        spans,
+    }
+}
+
+fn option_delta_i64(before: Option<u64>, after: Option<u64>) -> Option<i64> {
+    Some(after? as i64 - before? as i64)
+}
+
+fn option_delta_f64(before: Option<f64>, after: Option<f64>) -> Option<f64> {
+    Some(after? - before?)
+}
+
+fn option_percent_delta(before: Option<f64>, after: Option<f64>) -> Option<f64> {
+    let before = before?;
+    let after = after?;
+    if before.abs() < f64::EPSILON {
+        if after.abs() < f64::EPSILON {
+            Some(0.0)
+        } else {
+            None
+        }
+    } else {
+        Some(((after - before) / before) * 100.0)
+    }
+}
+
+pub(super) fn aggregate_span(
+    id: String,
+    samples: Vec<TraceAggregateSpanSample>,
+    failures: usize,
+) -> extension_trace::TraceAggregateSpanOutput {
+    let max_sample: Option<TraceAggregateSpanSample> =
+        samples.iter().fold(None, |max, sample| match max {
+            Some(current) if current.duration_ms >= sample.duration_ms => Some(current),
+            _ => Some(sample.clone()),
+        });
+    let mut durations = samples
+        .into_iter()
+        .map(|sample| sample.duration_ms)
+        .collect::<Vec<_>>();
+    durations.sort_unstable();
+    let n = durations.len();
+    let avg_ms = if n == 0 {
+        None
+    } else {
+        Some(durations.iter().sum::<u64>() as f64 / n as f64)
+    };
+    extension_trace::TraceAggregateSpanOutput {
+        id,
+        n,
+        min_ms: durations.first().copied(),
+        median_ms: median(&durations),
+        avg_ms,
+        p75_ms: percentile(&durations, 75, 4),
+        p90_ms: percentile(&durations, 90, 10),
+        p95_ms: percentile(&durations, 95, 20),
+        max_ms: durations.last().copied(),
+        max_run_index: max_sample.as_ref().map(|sample| sample.run_index),
+        max_artifact_path: max_sample.map(|sample| sample.artifact_path),
+        failures,
+    }
+}
+
+#[derive(Clone)]
+pub(super) struct TraceAggregateSpanSample {
+    pub(super) duration_ms: u64,
+    pub(super) run_index: usize,
+    pub(super) artifact_path: String,
+}
+
+fn median(values: &[u64]) -> Option<u64> {
+    if values.is_empty() {
+        return None;
+    }
+    let midpoint = values.len() / 2;
+    if values.len() % 2 == 1 {
+        Some(values[midpoint])
+    } else {
+        Some((values[midpoint - 1] + values[midpoint]) / 2)
+    }
+}
+
+fn percentile(values: &[u64], percentile: usize, min_samples: usize) -> Option<u64> {
+    if values.len() < min_samples {
+        return None;
+    }
+    let index = (values.len() * percentile).div_ceil(100).saturating_sub(1);
+    values.get(index).copied()
+}
+
+pub(super) fn render_aggregate_markdown(
+    aggregate: &extension_trace::TraceAggregateOutput,
+) -> String {
+    let mut out = String::new();
+    out.push_str(&format!(
+        "# Trace Aggregate: `{}`\n\n",
+        aggregate.scenario_id
+    ));
+    out.push_str(&format!("- **Component:** `{}`\n", aggregate.component));
+    out.push_str(&format!("- **Status:** `{}`\n", aggregate.status));
+    out.push_str(&format!("- **Runs:** `{}`\n", aggregate.run_count));
+    out.push_str(&format!("- **Failures:** `{}`\n", aggregate.failure_count));
+
+    if !aggregate.spans.is_empty() {
+        out.push_str("\n## Spans\n\n");
+        out.push_str("| Span | n | min | median | avg | p75 | p90 | p95 | max | failures |\n");
+        out.push_str("|---|---:|---:|---:|---:|---:|---:|---:|---:|---:|\n");
+        for span in &aggregate.spans {
+            out.push_str(&format!(
+                "| `{}` | {} | {} | {} | {} | {} | {} | {} | {} | {} |\n",
+                span.id,
+                span.n,
+                fmt_ms(span.min_ms),
+                fmt_ms(span.median_ms),
+                span.avg_ms
+                    .map(|value| format!("{:.1}ms", value))
+                    .unwrap_or_else(|| "-".to_string()),
+                fmt_ms(span.p75_ms),
+                fmt_ms(span.p90_ms),
+                fmt_ms(span.p95_ms),
+                fmt_ms(span.max_ms),
+                span.failures
+            ));
+        }
+
+        let outliers = aggregate
+            .spans
+            .iter()
+            .filter(|span| span.max_ms.is_some() && span.max_run_index.is_some())
+            .collect::<Vec<_>>();
+        if !outliers.is_empty() {
+            out.push_str("\n## Outliers\n\n");
+            for span in outliers {
+                out.push_str(&format!(
+                    "- `{}`: run {}, max={}, artifact=`{}`\n",
+                    span.id,
+                    span.max_run_index.unwrap_or_default(),
+                    fmt_ms(span.max_ms),
+                    span.max_artifact_path.as_deref().unwrap_or("")
+                ));
+            }
+        }
+    }
+
+    out.push_str("\n## Run Artifacts\n\n");
+    for run in &aggregate.runs {
+        out.push_str(&format!(
+            "- Run {}: `{}` `{}`\n",
+            run.index, run.status, run.artifact_path
+        ));
+    }
+    out
+}
+
+pub(super) fn render_compare_markdown(compare: &extension_trace::TraceCompareOutput) -> String {
+    let mut out = String::new();
+    out.push_str("# Trace Compare\n\n");
+    out.push_str(&format!("- **Before:** `{}`\n", compare.before_path));
+    out.push_str(&format!("- **After:** `{}`\n", compare.after_path));
+    if let (Some(before), Some(after)) = (&compare.before_scenario_id, &compare.after_scenario_id) {
+        out.push_str(&format!("- **Scenario:** `{}` -> `{}`\n", before, after));
+    }
+
+    if !compare.spans.is_empty() {
+        out.push_str("\n## Spans\n\n");
+        out.push_str("| Span | before median | after median | median delta | median % | before avg | after avg | avg delta | avg % |\n");
+        out.push_str("|---|---:|---:|---:|---:|---:|---:|---:|---:|\n");
+        for span in &compare.spans {
+            out.push_str(&format!(
+                "| `{}` | {} | {} | {} | {} | {} | {} | {} | {} |\n",
+                span.id,
+                fmt_ms(span.before_median_ms),
+                fmt_ms(span.after_median_ms),
+                fmt_delta_ms(span.median_delta_ms),
+                fmt_percent(span.median_delta_percent),
+                fmt_avg_ms(span.before_avg_ms),
+                fmt_avg_ms(span.after_avg_ms),
+                fmt_delta_avg_ms(span.avg_delta_ms),
+                fmt_percent(span.avg_delta_percent),
+            ));
+        }
+    }
+
+    out
+}
+
+fn fmt_ms(value: Option<u64>) -> String {
+    value
+        .map(|value| format!("{}ms", value))
+        .unwrap_or_else(|| "-".to_string())
+}
+
+fn fmt_avg_ms(value: Option<f64>) -> String {
+    value
+        .map(|value| format!("{:.1}ms", value))
+        .unwrap_or_else(|| "-".to_string())
+}
+
+fn fmt_delta_ms(value: Option<i64>) -> String {
+    value
+        .map(|value| format!("{:+}ms", value))
+        .unwrap_or_else(|| "-".to_string())
+}
+
+fn fmt_delta_avg_ms(value: Option<f64>) -> String {
+    value
+        .map(|value| format!("{:+.1}ms", value))
+        .unwrap_or_else(|| "-".to_string())
+}
+
+fn fmt_percent(value: Option<f64>) -> String {
+    value
+        .map(|value| format!("{:+.1}%", value))
+        .unwrap_or_else(|| "-".to_string())
+}
